requirements.txt
streamlit>=1.49,<1.50
SQLAlchemy>=2.0,<3.0
PyMySQL>=1.1,<2.0
pandas>=2.2,<3.0
numpy>=2.1,<3.0
altair>=5.3,<6.0
plotly>=5.24,<6.0
python-dotenv>=1.0,<2.0
requests>=2.32,<3.0
pycryptodome>=3.21,<4.0
tenacity>=8.2,<9.0
pytest>=8.3,<9.0
pytest-cov>=5.0,<6.0
black>=24.10,<25.0
isort>=5.13,<6.0
ruff>=0.6,<0.7
